allow import react components, not only renderer
