+++ package.json
@@ -1,15 +1,15 @@
 {
-  "name": "steal-jsx",
+  "name": "steal-react-jsx",
   "version": "0.0.2",
-  "description": "A StealJS plugin that enables importing renderers from `*.jsx` files.",
+  "description": "A StealJS plugin that enables importing reacts `*.jsx` files.",
   "main": "steal-jsx.js",
   "repository": {
     "type": "git",
-    "url": "git://github.com/stealjs/steal-jsx.git"
+    "url": "git://github.com/stealjs/steal-react-jsx.git"
   },
   "scripts": {
     "test": "npm run jshint && testee test/test.html --browsers firefox --reporter Spec",
-    "jshint": "jshint ./steal-jsx.js --config"
+    "jshint": "jshint ./steal-react-jsx.js --config"
   },
   "author": "Bitovi",
   "license": "MIT",
@@ -19,7 +19,7 @@
       "blacklist": []
     },
     "ext": {
-      "jsx": "steal-jsx"
+      "jsx": "steal-react-jsx"
     }
   },
   "devDependencies": {
