+++ test/templates/es6.jsx
