+++ test/templates/react-html.jsx
@@ -1,0 +1,4 @@
+
+
+
+		<div>Hello {ctx.props.name}!</div>
