--- test/templates/template.jsx
@@ -1,1 +1,0 @@
-<div>Hello {ctx.props.name}!</div>
